refactor: Delete confirm modal for components
