+++ src/library-authoring/LibraryAuthoringPage.scss
@@ -13,7 +13,7 @@
 
 .library-authoring-sidebar {
   z-index: 1000; // same as header
-  flex: 455px 0 0;
+  flex: 500px 0 0;
   position: sticky;
   top: 0;
   right: 0;
